backend/telegram_bot/requirements.txt
telegram_bot
├── src
│   ├── __init__.py
│   ├── main.py
│   ├── bot
│   │   ├── __init__.py
│   │   ├── trading_bot.py
│   │   ├── handlers
│   │   │   ├── __init__.py
│   │   │   ├── command_handlers.py
│   │   │   ├── callback_handlers.py
│   │   │   ├── message_handlers.py
│   │   │   └── error_handlers.py
│   │   ├── keyboards
│   │   │   ├── __init__.py
│   │   │   ├── main_keyboard.py
│   │   │   ├── analysis_keyboard.py
│   │   │   └── watchlist_keyboard.py
│   │   └── utils
│   │       ├── __init__.py
│   │       ├── state_manager.py
│   │       ├── message_formatter.py
│   │       └── validators.py
│   ├── services
│   │   ├── __init__.py
│   │   ├── watchlist_service.py
│   │   ├── analysis_service.py
│   │   └── scheduler_service.py
│   ├── models
│   │   ├── __init__.py
│   │   ├── user_state.py
│   │   └── watchlist_item.py
│   ├── data
│   │   ├── __init__.py
│   │   └── file_storage.py
│   └── config
│       ├── __init__.py
│       ├── settings.py
│       └── logging_config.py
├── requirements.txt
├── .env.example
├── .gitignore
└── README.md

requirements.txt
python-telegram-bot==20.0
apscheduler==3.8.1
pandas==1.3.3
numpy==1.21.2
asyncio==3.4.3
```